+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### v1.2.0-dev; 2025-09-13.2100
+### v1.2.0-dev; 2025-09-14.1600
 ```
 addressed raw base-16 issue https://github.com/cyclone-github/hashgen/issues/8
 added feature: "keep-order" from https://github.com/cyclone-github/hashgen/issues/7
@@ -7,6 +7,10 @@ add modes: mysql5 (300), phpass (400), md5crypt (500), sha256crypt (7400), sha51
 added hashcat salted modes: -m 10, 20, 110, 120, 1410, 1420, 1310, 1320, 1710, 1720, 10810, 10820
 added hashcat modes: -m 2600, 4500
 cleaned up hashFunc aliases, algo typo, and hex mode
+fixed ntlm encoding issue
+added sanity check to not print blank / invalid hashes (part of ntlm fix, but applies to all hash modes)
+converted checkForHex from string to byte
+updated yescrypt defaults to match debian 12 (libxcrypt)
 ```
 ### v1.1.4; 2025-08-23
 ```